implemented part of the rendering pipeline
now it renders the object with hard-coded MVP and flat shading with fixed color (defined in shader).
next step: transform system and camera
tentative plan is to make Transform a special attribute managed by EntityManager, while making camera a component that's attached to an entity.
That means to embark on the update loop messaging too.